+++ sdk/core/Azure.Core/src/Identity/ChainedTokenCredentialFactory.cs
@@ -315,11 +315,11 @@ namespace Azure.Identity
                 Pipeline = CredentialPipeline.GetInstance(managedIdentityOptions, IsManagedIdentityCredential: true),
                 Options = managedIdentityOptions,
                 IsForceRefreshEnabled = managedIdentityOptions.IsForceRefreshEnabled,
-                DisableMtlsProofOfPossession = managedIdentityOptions.DisableMtlsProofOfPossession,
+                EnableMtlsProofOfPossession = managedIdentityOptions.EnableMtlsProofOfPossession,
             }));
 
             var assertionOptions = source.Clone<ClientAssertionCredentialOptions>();
-            assertionOptions.DisableMtlsProofOfPossession = managedIdentityOptions.DisableMtlsProofOfPossession;
+            assertionOptions.EnableMtlsProofOfPossession = managedIdentityOptions.EnableMtlsProofOfPossession;
 
             if (source.AdditionallyAllowedTenants?.Count > 0)
             {