+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.63.0 (2026-09-23)
+## 1.62.1 (2026-09-23)
 
 ### Features Added